+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Changed
 - The Finalize Action now uses the SalesChannel from the Order. However, in some cases, the SalesChannel is not set correctly, which can result in incorrect API keys being used during the Finalize Action.
+### Fixed
+- Fixed the doctrine parameter types in elastic search and migrations
 
 ## [4.17.0] - 2025-08-04
 ### Added